Remove workaround for InsecureRequestWarning
Fixes #14

> from requests.packages.urllib3.exceptions import InsecureRequestWarning

This used to avoid telling the user that they were using an insecure version of Python.  This workaround is not longer supported because it is not good to ignore that the platform is insecure.  Repeated warnings will push people to upgrade to a secure version of Python. https://stackoverflow.com/questions/27981545/suppress-insecurerequestwarning-unverified-https-request-is-being-made-in-pytho